refactor: modularized transcription architecture and security fixes
