+++ .env.example
@@ -1,6 +1,7 @@
 HOST=127.0.0.1
 PORT=8000
 SECRET_KEY=change-me
+API_KEY_ENCRYPTION_KEY=change-this-encryption-key-in-production
 BOOTH_ACCESS_TOKEN=
 
 # JWT settings — leave JWT_SECRET empty to fall back to SECRET_KEY
