+++ alembic.ini
@@ -31,7 +31,7 @@ prepend_sys_path = .
 
 # set to 'true' to run the environment during
 # the 'revision' command, regardless of autogenerate
-revision_environment = true
+# revision_environment = false
 
 # set to 'true' to allow .pyc and .pyo files without
 # a source .py file to be detected as revisions in the
