+++ alembic/versions/008_add_and_encrypt_api_keys.py
@@ -1,0 +1,48 @@
+"""add_and_encrypt_api_keys
+
+Revision ID: 008
+Revises: 007
+Create Date: 2026-06-08 10:37:16.890313
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '008'
+down_revision: Union[str, Sequence[str], None] = '007'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    with op.batch_alter_table('booths') as batch_op:
+        batch_op.add_column(sa.Column('transcription_provider', sa.String(length=20), server_default=sa.text("'local'"), nullable=False))
+    
+    with op.batch_alter_table('events') as batch_op:
+        batch_op.add_column(sa.Column('transcription_api_enabled', sa.Boolean(), server_default='0', nullable=False))
+        batch_op.add_column(sa.Column('openai_api_key', sa.Text(), nullable=True))
+        batch_op.add_column(sa.Column('deepgram_api_key', sa.Text(), nullable=True))
+        batch_op.add_column(sa.Column('nvidia_api_key', sa.Text(), nullable=True))
+        batch_op.add_column(sa.Column('elevenlabs_api_key', sa.Text(), nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    with op.batch_alter_table('events') as batch_op:
+        batch_op.drop_column('elevenlabs_api_key')
+        batch_op.drop_column('nvidia_api_key')
+        batch_op.drop_column('deepgram_api_key')
+        batch_op.drop_column('openai_api_key')
+        batch_op.drop_column('transcription_api_enabled')
+        
+    with op.batch_alter_table('booths') as batch_op:
+        batch_op.drop_column('transcription_provider')
+    # ### end Alembic commands ###
