--- alembic/versions/008_add_transcription_api_keys.py
@@ -1,40 +1,0 @@
-"""add_transcription_api_keys
-
-Revision ID: 008
-Revises: 007
-Create Date: 2026-06-07 22:03:22.010336
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '008'
-down_revision: Union[str, Sequence[str], None] = '007'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('booths', sa.Column('transcription_provider', sa.String(length=20), server_default=sa.text("'local'"), nullable=False))
-    op.add_column('events', sa.Column('transcription_api_enabled', sa.Boolean(), server_default='0', nullable=False))
-    op.add_column('events', sa.Column('openai_api_key', sa.String(length=255), nullable=True))
-    op.add_column('events', sa.Column('deepgram_api_key', sa.String(length=255), nullable=True))
-    op.add_column('events', sa.Column('nvidia_api_key', sa.String(length=255), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('events', 'nvidia_api_key')
-    op.drop_column('events', 'deepgram_api_key')
-    op.drop_column('events', 'openai_api_key')
-    op.drop_column('events', 'transcription_api_enabled')
-    op.drop_column('booths', 'transcription_provider')
-    # ### end Alembic commands ###
