--- alembic/versions/009_add_elevenlabs_api_key.py
@@ -1,32 +1,0 @@
-"""add_elevenlabs_api_key
-
-Revision ID: 009
-Revises: 008
-Create Date: 2026-06-07 18:10:31.552047
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '009'
-down_revision: Union[str, Sequence[str], None] = '008'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('events', sa.Column('elevenlabs_api_key', sa.String(length=255), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('events', 'elevenlabs_api_key')
-    # ### end Alembic commands ###
