--- alembic/versions/010_encrypt_api_keys.py
@@ -1,38 +1,0 @@
-"""encrypt_api_keys
-
-Revision ID: 010
-Revises: 009
-Create Date: 2026-06-08 00:33:13.975815
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '010'
-down_revision: Union[str, Sequence[str], None] = '009'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table('events') as batch_op:
-        batch_op.alter_column('openai_api_key', type_=sa.Text())
-        batch_op.alter_column('deepgram_api_key', type_=sa.Text())
-        batch_op.alter_column('nvidia_api_key', type_=sa.Text())
-        batch_op.alter_column('elevenlabs_api_key', type_=sa.Text())
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table('events') as batch_op:
-        batch_op.alter_column('openai_api_key', type_=sa.String(255))
-        batch_op.alter_column('deepgram_api_key', type_=sa.String(255))
-        batch_op.alter_column('nvidia_api_key', type_=sa.String(255))
-        batch_op.alter_column('elevenlabs_api_key', type_=sa.String(255))
-    # ### end Alembic commands ###
