+++ portal/transcription/__init__.py
@@ -1,0 +1,19 @@
+from portal.transcription.constants import ProviderEnum, ALLOWED_MODELS
+from portal.transcription.providers.base import ProviderConfig, get_api_key
+from portal.transcription.worker import (
+    active_workers,
+    active_processes,
+    start_transcription_worker,
+    stop_transcription_worker,
+)
+
+__all__ = [
+    "ProviderEnum",
+    "ALLOWED_MODELS",
+    "ProviderConfig",
+    "get_api_key",
+    "active_workers",
+    "active_processes",
+    "start_transcription_worker",
+    "stop_transcription_worker",
+]
