+++ portal/transcription/constants.py
@@ -1,0 +1,16 @@
+from enum import Enum
+
+class ProviderEnum(str, Enum):
+    LOCAL = "local"
+    OPENAI = "openai"
+    DEEPGRAM = "deepgram"
+    NVIDIA = "nvidia"
+    ELEVENLABS = "elevenlabs"
+
+ALLOWED_MODELS = {
+    ProviderEnum.LOCAL: {"tiny", "base", "small", "medium", "large-v2", "large-v3"},
+    ProviderEnum.OPENAI: {"whisper-1", "gpt-4o-transcribe", "gpt-4o-mini-transcribe", "gpt-realtime-whisper"},
+    ProviderEnum.DEEPGRAM: {"nova-2"},
+    ProviderEnum.NVIDIA: {"parakeet-rnnt", "parakeet-ctc"},
+    ProviderEnum.ELEVENLABS: {"scribe_v2"}
+}
